Mitm/src/requirements.txt
scapy==2.6.1
termcolor==2.5.0
PyYAML==6.0.2